% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{align_columns}
\alias{align_columns}
\title{Align Columns}
\usage{
align_columns(x, method = c("fixed_width", "automatic"), ...)
}
\arguments{
\item{x}{a object inheriting from \code{'data.frame'}.}

\item{method}{a character string giving the method.}

\item{...}{additional arguments}
}
\value{
Returns an object inheriting from \code{'data.frame'}.
}
\description{
TODO
}
