% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_lines.R
\name{align_hlines}
\alias{align_hlines}
\title{Align horizontal lines}
\usage{
align_hlines(x, tol = 1)
}
\arguments{
\item{x}{an object of class dataframe.}

\item{tol}{a threshold of how many distance can be between line segments
which should still be handled as connected.}

\item{...}{additional arguments}
}
\value{
Returns a df where all horizontal line segments
with connecting end and start point are concatened.
}
\description{
Align the horizontal lines of a dataframe.
}
\details{
Some details to be written.
}
