% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_lines}
\alias{align_lines}
\title{Align all lines}
\usage{
align_lines(x)
}
\arguments{
\item{x}{an object of class dataframe.}

\item{tol}{a threshold of how many distance can be between line segments
which should still be handled as connected.}

\item{...}{additional arguments}
}
\value{
Returns a df where all line segments (which belong to the same
horizontal or vertical) with connecting end and start point are concatened.
}
\description{
Align the all (horizontal + vertical) lines of a dataframe.
}
\details{
Some details to be written.
}
