% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_lines.R
\name{align_vlines}
\alias{align_vlines}
\title{Align vertical lines}
\usage{
align_vlines(x, tol = 1, ...)
}
\arguments{
\item{x}{an object of class dataframe.}

\item{tol}{a threshold of how many distance can be between line segments
which should still be handled as connected.}

\item{...}{additional arguments (currently not used).}
}
\value{
Returns a df where all vertical line segments
with connecting end and start point are concatened.
}
\description{
Align the vertical lines of a dataframe.
}
\details{
Some details to be written.
}
