% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bboxplot}
\alias{bboxplot}
\title{bboxplot}
\usage{
bboxplot(x, split_points = NULL, pid = 1L, ...)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{split_points}{an vector containing the x-coordinates for splitting
the page.}

\item{pid}{the number of the page which shall be plotted.}

\item{...}{additional arguments (currently not used).}
}
\value{
TODO
}
\description{
Plot the detected bounding boxes
}
\details{
Some details to be written.
}
\examples{
pdf_file <- system.file(file.path("pdfs","cars.pdf"), package = "pdfmole")
pdf <- pdfminer::read.pdf(pdf_file, pages = 1:2, maxpages = 2L)

df <- pdf$text
head(df, 20)

bboxplot(df, pid = 1L)
}
