% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorplot.R
\name{colorplot}
\alias{colorplot}
\title{colorplot}
\usage{
colorplot(x, split_points = NULL, pid = 0)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{split_points}{an vector containing the x-coordinates for splitting
the page.}

\item{pid}{the number of the page which shall be plotted.}

\item{...}{additional arguments (currently not used).}
}
\value{
TODO
}
\description{
TODO
}
\details{
Some details to be written.
}
