% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{data_pdfminer}
\alias{data_pdfminer}
\title{Data Obtained by \code{pdfminer::read_chars}}
\format{
an object of class \code{pdf\_document} obtained by \code{pdfminer::read_chars}.

an object of class \code{pdf\_document} obtained by \code{pdfboxr::read_chars}.
}
\description{
This dataset contains the data from \code{"cars.pdf"} obtained 
by \code{pdfminer::read_chars}.

This dataset contains the data from \code{"cars.pdf"} obtained 
by \code{pdfminer::read_chars}.
}
\keyword{datasets}
