% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{data_tesseract}
\alias{data_tesseract}
\title{Data Obtained by \code{tesseract::ocr_data}}
\format{
A tibble as obtained by \code{tesseract::ocr_data}.
}
\description{
This dataset contains the first page of \code{"cars.pdf"} obtained 
by \code{tesseract::ocr_data}.
}
\keyword{datasets}
