% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{distribute_breaks}
\alias{distribute_breaks}
\title{Distributes breaks}
\usage{
distribute_breaks(x, number_cols)
}
\arguments{
\item{x}{an object inheriting from \code{'data.frame'}.}

\item{number_cols}{the nubmer of columns of the document.}
}
\value{
Returns an vector containing the column breaks.
}
\description{
Distributes equidistantly the given number of columns.
}
