% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{find_breaks}
\alias{find_breaks}
\title{Find Breaks}
\usage{
find_breaks(x)
}
\arguments{
\item{x}{a object inheriting from \code{'data.frame'}.}
}
\value{
Returns an vector containing the colum breaks.
}
\description{
Find Breaks
}
