% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesseract.R
\name{from_tesseract}
\alias{from_tesseract}
\title{Convert Data from Tesseract}
\usage{
from_tesseract(x)
}
\arguments{
\item{x}{a tibble obtained by \code{tesseract::ocr_data}.}
}
\value{
a \code{data.frame} usable by \pkg{pdfmole}.
}
\description{
Convert data obtained by \code{tesseract::ocr_data} into
an format usable by \pkg{pdfmole}.
}
\examples{
# ocr_data <- tesseract::ocr_data("my_file".png)
# data("ocr_data")
# add page number
# ocr_data$pid <- 1L
# from_tesseract(ocr_data)
}
