% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_blocks}
\alias{group_blocks}
\title{Group Blocks}
\usage{
group_blocks(x, collapse = "")
}
\arguments{
\item{x}{an object inheriting from \code{'data.frame'}.}

\item{collapse}{a character string to separate the results.}
}
\value{
Returns an object inheriting from \code{'data.frame'}.
}
\description{
Groups cells based on the determined block value
}
\details{
This is an optional step...
}
