% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_columns}
\alias{group_columns}
\title{Group Columns}
\usage{
group_columns(x, collapse = "")
}
\arguments{
\item{x}{an object inheriting from \code{'data.frame'}.}

\item{collapse}{a character string to separate the results.}
}
\value{
Returns an object inheriting from \code{'data.frame'}.
}
\description{
Groups cells based on the determined page, row and col values.
}
\details{
TODO
}
