% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pdfminer.R
\name{install_pdfminer}
\alias{install_pdfminer}
\title{Install \pkg{pdfminer}}
\usage{
install_pdfminer(pdfminer_dir, user = "pdfminer",
  repo = "pdfminer.six", ...)
}
\arguments{
\item{pdfminer_dir}{a character string giving the path where \pkg{pdfminer} 
should be installed.}

\item{user}{a character string giving the user name of the \pkg{pdfminer} 
github repository owner, default is \code{"pdfminer"}.}

\item{repo}{a character string giving the name of the \pkg{pdfminer} github 
repository, default is \code{"pdfminer.six"}.}

\item{...}{additional arguments passed to \code{download.file}.}
}
\description{
Install pdfminer from \R.
}
\details{
To use \pkg{pdfmole} the \code{Python} package \pkg{pdfminer}
  needs to be available. There exist two versions of \pkg{pdfminer},
  \pkg{pdfminer} and \pkg{pdfminer.six}, where \pkg{pdfminer.six} is a fork
  from \pkg{pdfminer}. For more information can be found in the \pkg{pdfmole}
  \code{README} file.
}
\examples{
\dontrun{
install_pdfminer("~/Software/pdfminer")
}
}
