% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalplot.R
\name{intervalplot}
\alias{intervalplot}
\title{Intervalplot}
\usage{
intervalplot(
  x,
  breaks = NULL,
  by = 0.1,
  offset = NULL,
  num_breaks = 500,
  grid_len = 20,
  widths = NULL,
  ...
)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{breaks}{TODO}

\item{by}{TODO}

\item{offset}{TODO}

\item{num_breaks}{TODO}

\item{grid_len}{TODO}

\item{widths}{TODO}

\item{...}{additional arguments (currently not used).}
}
\value{
TODO
}
\description{
TODO
}
\details{
Some details to be written.
}
