% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfmole.R
\name{mole}
\alias{mole}
\title{Convert into Rectangular Format.}
\usage{
mole(x, header = FALSE, simplify = FALSE, keep = FALSE)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{header}{a logical indicating whether the first line contains the names 
of the variables.}

\item{simplify}{a logical giving if the results should be simplified.
If \code{FALSE} all vectors are character vectors, if \code{TRUE}
the correct types are inferred by \pkg{pdfmole}.}

\item{keep}{a logical giving if the variables \code{"pid"} and \code{"row"}
should be kept in the output.}
}
\value{
Returns an object of class \code{"mole"}.
}
\description{
Convert the data from a sparse sparse simple triplet representation
 into a dense list of vectors format.
}
