% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pdfminer.R
\name{pdfminer_set_environment_variable}
\alias{pdfminer_set_environment_variable}
\title{Set \pkg{pdfminer} Environment Variable}
\usage{
pdfminer_set_environment_variable(pdfminer_dir)
}
\arguments{
\item{pdfminer_dir}{a character string giving the path where \pkg{pdfminer} 
should be installed.}
}
\description{
Set the environment variable \code{"PDFMINER\_HOME"}.
}
\details{
The environment variable \code{"PDFMINER\_HOME"} contains the 
  path to \pkg{pdfminer}, this information if needed in order to load
  \pkg{pdfminer} from \pkg{pdfmole}.
}
\examples{
\dontrun{
pdfminer_set_environment_variable("~/Software/pdfminer")
}
}
