% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pixelplot}
\alias{pixelplot}
\title{Pixelplot}
\usage{
pixelplot(x, scale = 1, pids = integer(), las = 2, cex.axis = 0.7, ...)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{scale}{a double giving the scale factor.}

\item{pids}{an integer vector giving the pages to be considered in the pixelplot.}

\item{las}{labels are parallel (=0) or perpendicular(=2) to axis}

\item{cex.axis}{size of axis annotation relative to cex}

\item{...}{additional arguments (currently not used).}
}
\description{
Plot the count of letters
}
\details{
Some details to be written.
}
\examples{
# df <- pdf$text
# head(df, 20)
# pixelplot(df, scale = 0.5)
}
