% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfmole.R
\name{read.pdf}
\alias{read.pdf}
\title{Read a \code{PDF} document.}
\usage{
read.pdf(file, pages = integer(), encoding = "utf8", maxpages = Inf,
  check_extractable = TRUE, password = "", caching = TRUE,
  layout_page_number = 1L)
}
\arguments{
\item{file}{a character giving the name of the \code{PDF}-file the data are to be read from.}

\item{pages}{an integer giving the pages pages which should be extracted.}

\item{encoding}{a character string giving the encoding of the output.}

\item{maxpages}{an integer giving the maximum number of pages to be extracted.}

\item{check_extractable}{a logical indicating if a check should be performend
to verify that the text is extractable.}

\item{password}{a character string giving the password necessary to access the \code{PDF}.}

\item{caching}{a logical if \code{TRUE} \pkg{pdfmole} is faster but uses more memory.}

\item{layout_page_number}{an integer giving number of the page where the layout
analysis should be performed.}
}
\value{
Returns a object of class \code{"pdf_document"}.
        A object of class \code{"pdf_document"} is a list where
        each element of the list is an object of class \code{"pdf_page"}.
}
\description{
Extract  \code{PDF} document
}
