% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_area.R
\name{select_area}
\alias{select_area}
\title{select_area}
\usage{
select_area(x, ...)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{...}{additional arguments (currently not used).}
}
\value{
TODO
}
\description{
A function creating a shiny miniUI for select the coordinates
for \code{\link{select_bounds}}.
}
\details{
Some details to be written.
}
