% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\name{textplot}
\alias{textplot}
\title{textplot}
\usage{
textplot(x, split_points = NULL, pid = 0L)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{split_points}{an vector containing the x-coordinates for splitting the columns.}

\item{pid}{the number of the page which shall be plotted.}

\item{...}{additional arguments (currently not used).}
}
\value{
TODO
}
\description{
TODO
}
\details{
Some details to be written.
}
