% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_rows}
\alias{align_rows}
\title{Align Rows}
\usage{
align_rows(x, method = c("exact_match", "hclust", "fixed_width"), ...)
}
\arguments{
\item{x}{an object of class \code{pdf\_df}.}

\item{method}{a character string giving the name of the method.}

\item{...}{additional arguments}
}
\value{
Returns a object of class \code{"pdf\_df"}.
}
\description{
Align the rows of an object of class \code{pdf\_df}.
}
\details{
Some details to be written.
}
