% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{find_breaks}
\alias{find_breaks}
\title{Find Breaks}
\usage{
find_breaks(x, lower_bound = 0.25)
}
\arguments{
\item{x}{an object inheriting from \code{'data.frame'}.}

\item{lower_bound}{a lower bound indicating with which ...}
}
\value{
Returns an vector containing the column breaks.
}
\description{
Find Breaks
}
