% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_blocks}
\alias{group_blocks}
\title{Group Blocks}
\usage{
group_blocks(x, collapse = "")
}
\arguments{
\item{x}{an object inheriting from \code{'data.frame'}.}
}
\value{
Returns an object inheriting from \code{'data.frame'}.
}
\description{
TODO
}
