% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfmole.R
\name{mole}
\alias{mole}
\title{Convert into Rectangular Format.}
\usage{
mole(x, header = FALSE, simplify = FALSE, keep = FALSE)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{header}{a logical}

\item{simplify}{a logical}

\item{keep}{a logical}
}
\description{
TODO
}
