% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{textplot}
\alias{textplot}
\title{textplot}
\usage{
textplot(
  x,
  split_points = NULL,
  pid = 1L,
  cex.text = 0.8,
  cex.xaxis = 0.8,
  ...
)
}
\arguments{
\item{x}{an object inheriting from \code{data.frame}.}

\item{split_points}{an vector containing the x-coordinates for splitting
the page.}

\item{pid}{the number of the page which shall be plotted.}

\item{cex.text}{a numerical value giving the amount by which the text in the plot
should be magnified relative to the default.}

\item{cex.xaxis}{a numerical value giving the amount by which the text on the x-axis
should be magnified relative to the default.}

\item{...}{additional arguments (currently not used).}
}
\description{
Plot the detected text
}
\details{
Some details to be written.
}
\examples{
# df <- pdf$text
# head(df, 20)
# textplot(df, pid = 1L)
}
